+++ src/scandeval/dataset_configs.py
@@ -938,6 +938,35 @@ CNN_DAILYMAIL_CONFIG = DatasetConfig(
     max_generated_tokens=256,
 )
 
+SCHIBSTED_SV = DatasetConfig(
+    name="schibsted-sv",
+    pretty_name="article summaries from Schibsted Media Swedish newsrooms.",
+    huggingface_id="ScandEval/schibsted-article-summaries-sv",
+    task=SUMM,
+    languages=[SV],
+    prompt_prefix="Nedan följer artiklar med tillhörande sammanfattningar.",
+    prompt_template="Artikel: {text}\nSammanfattning: {target_text}",
+    instruction_prompt="Artikel: {text}\n\nSkriv en sammanfattning av artikeln ovan.",
+    num_few_shot_examples=1,
+    max_generated_tokens=256,
+    unofficial=True,
+)
+
+SCHIBSTED_NO = DatasetConfig(
+    name="schibsted-no",
+    pretty_name="article summaries from Schibsted Medias Norwegian newsrooms.",
+    huggingface_id="ScandEval/schibsted-article-summaries-no",
+    task=SUMM,
+    languages=[NB, NN, NO],
+    prompt_prefix="Her følger nyhetsartikler med tilhørende sammendrag.",
+    prompt_template="Nyhetsartikkel: {text}\nSammendrag: {target_text}",
+    instruction_prompt="Nyhetsartikkel: {text}\n\nSkriv et sammendrag av den "
+    "ovennevnte artikkelen.",
+    num_few_shot_examples=1,
+    max_generated_tokens=256,
+    unofficial=True,
+)
+
 # TODO: Faroese summarization