+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Added a linguistic acceptability dataset which is a subset of the [Icelandic Linguistic Benchmarks dataset](https://raw.githubusercontent.com/stofnun-arna-magnussonar/ice_linguistic_benchmarks/refs/heads/main/ice_benchmark_set.jsonl). It is a small dataset with 250 train samples, 50 val samples, and 82 test samples. The dataset has been marked as unofficial.
 - Added the [Schibsted
   dataset](https://huggingface.co/datasets/Schibsted/schibsted-article-summaries), which
   contains summaries of published articles from Schibsted Media's Norwegian and Swedish